Modificando controllers (security e service)

+++ BlogPessoal/src/main/java/com/blogPessoal/BlogPessoal/controllers/PostagemController.java
@@ -1,6 +1,9 @@
 package com.blogPessoal.BlogPessoal.controllers;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,51 +16,57 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.blogPessoal.BlogPessoal.models.Postagem;
 import com.blogPessoal.BlogPessoal.repositories.PostagemRepository;
 
 @RestController
-@RequestMapping ("/postagens")
-@CrossOrigin ("*")
+@RequestMapping("/postagens")
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 public class PostagemController {
 
 	@Autowired
 	private PostagemRepository repository;
-	
+
 	@GetMapping
-	public ResponseEntity <List <Postagem>> findAllPostagem() {
+	public ResponseEntity<List<Postagem>> findAllPostagem() {
 		return ResponseEntity.ok(repository.findAll());
 	}
-	
-	@GetMapping ("/{id}")
-	public ResponseEntity <Postagem> findByIDPostagem (@PathVariable long id) {
-		return repository.findById(id)
-				.map(resp -> ResponseEntity.ok(resp))
-				.orElse(ResponseEntity.notFound().build());
+
+	@GetMapping("/{id}")
+	public ResponseEntity<Postagem> findByIDPostagem(@PathVariable long id) {
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
-	
+
 	@GetMapping("/titulo/{titulo}")
-	public ResponseEntity<List<Postagem>> findByDescricaoTitulo (@PathVariable String titulo) {
+	public ResponseEntity<List<Postagem>> findByDescricaoTitulo(@PathVariable String titulo) {
 		return ResponseEntity.ok(repository.getByTituloContainingIgnoreCase(titulo));
 	}
-	
+
 	@PostMapping
-	public ResponseEntity <Postagem> postPostagem (@RequestBody Postagem postagem) {
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(repository.save(postagem));
+	public ResponseEntity<Postagem> postPostagem(@Valid @RequestBody Postagem postagem) {
+		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(postagem));
 	}
-	
+
 	@PutMapping
-	public ResponseEntity <Postagem> putPostagem (@RequestBody Postagem postagem) {
-		return ResponseEntity.status(HttpStatus.OK)
-				.body(repository.save(postagem));
+	public ResponseEntity<Postagem> putPostagem(@Valid @RequestBody Postagem postagem) {
+		return repository.findById(postagem.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.OK).body(repository.save(postagem)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
-	
-	@DeleteMapping ("/{id}")
-	public void deletePostagem (@PathVariable long id) {
+
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void deletePostagem(@PathVariable long id) {
+		Optional<Postagem> post = repository.findById(id);
+		if (post.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repository.deleteById(id);
 	}
-	
+
 }
